remove pipe_plugin
