+++ pytorch_lightning/core/optimizer.py
@@ -214,8 +214,9 @@ class LightningOptimizer:
                     xm.optimizer_step(optimizer, optimizer_args={'closure': closure, **kwargs})
 
             elif trainer.amp_backend is not None:
-                trainer.precision_connector.backend.optimizer_step(
-                    trainer, optimizer, closure)
+                with trainer.profiler.profile(profiler_name):
+                    trainer.precision_connector.backend.optimizer_step(
+                        trainer, optimizer, closure)
 
             else:
                 with trainer.profiler.profile(profiler_name):
